Some binary operator tests.

+++ ir/value/enum_and_flags.h
@@ -5,8 +5,8 @@
 
 namespace ir {
 
-ICARUS_BASE_DEFINE_STRONG_TYPE(EnumVal, size_t{0}, base::EnableHashing);
-ICARUS_BASE_DEFINE_STRONG_TYPE(FlagsVal, size_t{0}, base::EnableHashing);
+ICARUS_BASE_DEFINE_STRONG_TYPE(EnumVal, uint64_t{0}, base::EnableHashing);
+ICARUS_BASE_DEFINE_STRONG_TYPE(FlagsVal, uint64_t{0}, base::EnableHashing);
 
 constexpr FlagsVal operator|(FlagsVal lhs, FlagsVal rhs) {
   return FlagsVal{lhs.value | rhs.value};
